Updates code

+++ Barrios/Barrios.Web/Modules/Default/Reservas/ReservasEndpoint.cs
@@ -20,14 +20,22 @@ namespace Barrios.Default.Endpoints
     using MyRow = Entities.ReservasRow;
     using Barrios.Modules.Default.Reservas;
     using Barrios.Administration.Entities;
+    using Barrios.Default.Repositories;
 
     [RoutePrefix("Services/Default/Reservas"), Route("{action}")]
     [ConnectionKey(typeof(MyRow)), ServiceAuthorize(typeof(MyRow))]
     public class ReservasController : ServiceEndpoint
     {
+        private void IdTurnsEspecialToIdType(MyRow obj)
+        {
+            ReservasRecursosRow resource = Utils.GetConnection().Query<ReservasRecursosRow>("SELECT * FROM [RESERVAS_RECURSOS] WHERE id=" + obj.IdRecurso).SingleOrDefault();
+            if (resource.Resolucion == 0)
+                obj.IdTipo = obj.IdTurnosEspeciales;
+        }
         [HttpPost, AuthorizeCreate(typeof(MyRow))]
         public SaveResponse Create(IUnitOfWork uow, SaveRequest<MyRow> request)
         {
+            IdTurnsEspecialToIdType(request.Entity);
             request.Entity.BarrioId = CurrentNeigborhood.Get().Id;
             var response= new MyRepository().Create(uow, request);
             uow.OnCommit += () =>
@@ -40,6 +48,7 @@ namespace Barrios.Default.Endpoints
         [HttpPost, AuthorizeUpdate(typeof(MyRow))]
         public SaveResponse Update(IUnitOfWork uow, SaveRequest<MyRow> request)
         {
+            IdTurnsEspecialToIdType(request.Entity);
             uow.OnCommit += () =>
             {
                 SendBookingTakeMail(null, request.Entity);
@@ -111,18 +120,12 @@ namespace Barrios.Default.Endpoints
 
 
             UserRow user = Utils.GetUser(Convert.ToInt32(Authorization.UserId));
-            bool ok = Convert.ToBoolean(Utils.GetRequestString( "SELECT CAST(CASE WHEN DATEDIFF(D, (SELECT TOP 1 FECHA FROM HOY), " + date.ToSql() + ") <= 180 THEN 1 ELSE 0 END AS BIT)").Rows[0][0]);
+            bool ok = Convert.ToBoolean(Utils.GetRequestString( "SELECT CAST(CASE WHEN DATEDIFF(D, (SELECT TOP 1 FECHA FROM HOY), " + date.ToSql() + ") <= 180 AND DATEDIFF(D, (SELECT TOP 1 FECHA FROM HOY), " + date.ToSql() + ") > 9 THEN 1 ELSE 0 END AS BIT)").Rows[0][0]);
             if (ok) {
                 string status = Utils.GetRequestString( "SELECT dbo.ESTADO_TURNO_RESERVA(" + request.resourceId + ", " + date.ToSql() + ", " + request.turnStart + ", " + request.turnDuration + ")").Rows[0][0].ToString();
 
                 if (status.ToUpper() == "DISPONIBLE") {
-                    MyRow obj = new MyRow
-                    {
-                        IdRecursoNombre = request.resourceName,
-                        Turno = request.turnName,
-                        Observaciones = request.comment,
-                        Fecha = date
-                    };
+                    MyRow obj = EmailHelper.GetReservaRowForBody(request, user, date, request.resourceName);
                     string emails =Utils.GetRequestString( "SELECT emails FROM [RESERVAS_RECURSOS] WHERE id=" + request.resourceId).Rows[0][0].ToString();
                     emails = EmailHelper.GetRenderMails(emails, user);
                     
@@ -141,20 +144,46 @@ namespace Barrios.Default.Endpoints
                 }
             }
             else {
-                throw new Exception("Las reservas del " + request.resourceName+ " deben solicitarse con una anticipación máxima de 6 meses.");
+                throw new Exception("Las reservas del " + request.resourceName+ " deben solicitarse con una anticipación minima de 10 dias y máxima de 6 meses.");
             }
             return "Se ha enviado la solicitud correctamente";
         }
         [HttpPost]
         public string bookingCancel(IDbConnection connection, IdRequest request)
         {
-            
-            var UOW = Utils.GetUnitOfWork();
-            Delete(UOW, new DeleteRequest() { EntityId = request.ID });
-            UOW.Commit();
+            if (ValidateForDelete(request.ID))
+            {
+                var UOW = Utils.GetUnitOfWork();
+                Delete(UOW, new DeleteRequest() { EntityId = request.ID });
+                UOW.Commit();
+            }
             return "Se ha eliminado su reserva.";
         }
-
+        private bool ValidateForDelete(int? id)
+        {
+            using (var connection1 = Utils.GetConnection())
+            {
+                MyRow row = Retrieve(connection1, new RetrieveRequest() { EntityId = id }).Entity;
+                using (var connection2 = Utils.GetConnection())
+                {
+                    ReservasRecursosRow resource = connection2.Query<ReservasRecursosRow>("SELECT * FROM [RESERVAS_RECURSOS] WHERE id=" + row.IdRecurso).SingleOrDefault();
+                    if (resource.Resolucion == 0)
+                    {
+                        if (row.Fecha < DateTime.Today.AddDays(10))
+                            throw new Exception("No puede cancelar su reserva con menos de 10 dias de anticipación. Comuniquese con admnistración.");
+                    }
+                    else
+                    {
+                        DataRow DR = Utils.GetRequestString("SELECT RR.NOMBRE,RR.RESOLUCION,R.ID_RECURSO,R.FECHA, dbo.HOUR_TO_STR(INICIO) AS TURNO, " +
+                     "CAST(CASE WHEN DATEDIFF(MINUTE, GETDATE(), dbo.FECHA_INICIO_TURNO(FECHA, INICIO)) >= 1 THEN 1 ELSE 0 END AS BIT) AS OK " +
+                     "FROM RESERVAS R INNER JOIN [RESERVAS_RECURSOS] RR ON R.ID_RECURSO=RR.ID WHERE R.ID=" + id).Rows[0];
+                        if (!Convert.ToBoolean(DR["OK"]))
+                            throw new ValidationError(DR["NOMBRE"] + ": el turno " + DR["FECHA"] + " " + DR["TURNO"] + "\n\nYa no es posible cancelar la reserva porque la anticipación mínima es de 1 minuto antes.");
+                    }
+                }
+                return true;
+            } 
+         }
         #region SEND MAILS
         private void SendBookingCancelMail(DeleteRequest request)
         {
@@ -162,38 +191,24 @@ namespace Barrios.Default.Endpoints
             {
                 MyRow row = Retrieve(connection1, new RetrieveRequest() { EntityId = request.EntityId }).Entity;
 
-                ReservasRecursosRow resource = Utils.GetConnection().Query<ReservasRecursosRow>("SELECT * FROM [RESERVAS_RECURSOS] WHERE id=" + row.IdRecurso).SingleOrDefault();
 
+                ReservasRecursosRow resource = Utils.GetConnection().Query<ReservasRecursosRow>("SELECT * FROM [RESERVAS_RECURSOS] WHERE id=" + row.IdRecurso).SingleOrDefault();
 
                 if (resource.Resolucion == 0)
                 {
-                    if (row.Fecha < DateTime.Today.AddDays(10))
+                    UserRow user = Utils.GetUser(row.IdVecino);
+                    UserRow user2 = Utils.GetUser(row.IdVecino2);
+                    string emails = EmailHelper.GetRenderMails(resource.Emails, user, user2);
+                    row.Turno = row.Tipo;
+                    row.IdVecinoUnidad = user.Unit;
+                    var message = TemplateHelper.RenderTemplate(MVC.Views.Default.Reservas.Mail.BookingEmail, new MailBody()
                     {
-                        throw new Exception("No puede cancelar su reserva con menos de 10 dias de anticipación. Comuniquese con admnistración.");
-                    }
-                    else
-                    {
-                        UserRow user = Utils.GetUser(row.IdVecino);
-                        UserRow user2 = Utils.GetUser(row.IdVecino2);
-                        string emails = EmailHelper.GetRenderMails(resource.Emails, user, user2);
-                        var message = TemplateHelper.RenderTemplate(MVC.Views.Default.Reservas.Mail.BookingEmail, new MailBody()
-                        {
-                            Reserva = row,
-                            Body = "",
-                            Title = "Reserva cancelada",
-                            BeforeTable = "A partir de este correo.Queda cancelada su reserva."
-                        });
-                        EmailHelper.Send($"Reserva {resource.Nombre} cancelada", message, emails, CurrentNeigborhood.Get().LargeDisplayName, CurrentNeigborhood.Get().Mail, null, resource.Regulation);
-
-                    }
-                }
-                else
-                {
-                    DataRow DR = Utils.GetRequestString("SELECT RR.NOMBRE,RR.RESOLUCION,R.ID_RECURSO,R.FECHA, dbo.HOUR_TO_STR(INICIO) AS TURNO, " +
-                 "CAST(CASE WHEN DATEDIFF(MINUTE, GETDATE(), dbo.FECHA_INICIO_TURNO(FECHA, INICIO)) >= 1 THEN 1 ELSE 0 END AS BIT) AS OK " +
-                 "FROM RESERVAS R INNER JOIN [RESERVAS_RECURSOS] RR ON R.ID_RECURSO=RR.ID WHERE R.ID=" + request.EntityId).Rows[0];
-                    if (!Convert.ToBoolean(DR["OK"]))
-                        throw new ValidationError(DR["NOMBRE"] + ": el turno " + DR["FECHA"] + " " + DR["TURNO"] + "\n\nYa no es posible cancelar la reserva porque la anticipación mínima es de 1 minuto antes.");
+                        Reserva = row,
+                        Body = "",
+                        Title = "Reserva cancelada",
+                        BeforeTable = "A partir de este correo.Queda cancelada su reserva."
+                    });
+                    EmailHelper.Send($"Reserva {resource.Nombre} cancelada", message, emails, CurrentNeigborhood.Get().LargeDisplayName, CurrentNeigborhood.Get().Mail, null, resource.Regulation);
                 }
             }
         }
@@ -201,26 +216,24 @@ namespace Barrios.Default.Endpoints
         {
             int resourceID = (request == null) ? row.IdRecurso.Value : request.resourceId;
             ReservasRecursosRow resource = Utils.GetConnection().Query<ReservasRecursosRow>("SELECT * FROM [RESERVAS_RECURSOS] WHERE id=" + resourceID).SingleOrDefault();
+            
             if (!resource.Emails.IsEmptyOrNull() || resource.Resolucion == 0)
             {
+                UserRow user=Utils.GetUser(row.IdVecino);
                 if (row == null)
                 {
                     DateTime date = DateTime.ParseExact(request.bookingDate, "yyyyMMdd",
                                       CultureInfo.InvariantCulture);
-                    row = new MyRow
-                    {
-                        IdRecursoNombre = resource.Nombre,
-                        Turno = request.turnStart.MinutesToString(),
-                        Observaciones = request.comment,
-                        Fecha = date
-                    };
+                    row = EmailHelper.GetReservaRowForBody(request, user, date, resource.Description);
                 }
                 else
                 {
-                    row.Turno = row.Inicio.Value.MinutesToString();
+                    row.Hora = row.Inicio.Value.MinutesToString();
+                    row.IdVecinoUnidad = user.Unit;
+                    row.IdVecinoUsername = user.DisplayName;
                     row.IdRecursoNombre = resource.Nombre;
                 }
-                string emails = EmailHelper.GetRenderMails(resource.Emails, Utils.GetUser(row.IdVecino), Utils.GetUser(row.IdVecino2));
+                string emails = EmailHelper.GetRenderMails(resource.Emails, user, Utils.GetUser(row.IdVecino2));
                 var message = TemplateHelper.RenderTemplate(MVC.Views.Default.Reservas.Mail.BookingEmail, new MailBody()
                 {
                     Reserva = row,
